Add data path selection
This update adds a new dialog to allow users to select the folder that
contains the country-specific data folders, rather than assuming the
folders exist in a subdirectory of the current working path. Users click
on the 'Select folder' button and choose the path to a folder that
contains subdirectories for each country (currently only Fiji and
Tonga). The dialog automatically updates the list of available projects
for that country, based on the QGIS project files in the subdirectory.
If the selected folder does not contain the expected directory strcture,
a message box is displayed, telling the user to select a directory where
the country projects are stored.